Updated version for new MSM API
