+++ my_school_menus/__init__.py
@@ -1,5 +1,1 @@
 from .__meta__ import __version__
-
-__all__ = (
-    "__version__"
-)
