+++ my_school_menus/__meta__.py
@@ -1,1 +1,1 @@
-__version__ = "0.0.9"
+__version__ = "0.1.0"
